Create Chartboost singleton delegate to be shared by both rewarded video and interstitial video.
